requirements.txt
python-dotenv>=1.0.1
tqdm>=4.66.4
pydantic>=2.9.0
tenacity>=8.5.0
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.4
openai>=1.43.0
anthropic>=0.34.2
google-generativeai>=0.8.0
